Refactored AppUser into identity folder in domain layer
