+++ Delpin.API/Controllers/v1/PostalCitiesController.cs
@@ -6,7 +6,6 @@ using Microsoft.Extensions.Logging;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
 
 namespace Delpin.API.Controllers.v1
 {
@@ -23,7 +22,7 @@ namespace Delpin.API.Controllers.v1
             _context = context;
             _logger = logger;
         }
-        
+
         [HttpGet]
         public async Task<ActionResult<IReadOnlyList<PostalCity>>> GetAll()
         {
